2/ refactor rout and repository delete photo

+++ src/routes/photos.py
@@ -9,7 +9,7 @@ from fastapi.security import (
     HTTPBearer,
 )
 from src.conf.config import settings
-from src.database.models import User
+from src.database.models import User, Role
 from src.repository import users as repository_users
 from src.repository import photos as repository_photos
 from src.database.db import get_db
@@ -43,14 +43,28 @@ async def create_photo(description: Optional[str] = None, db: Session = Depends(
 async def delete_photo(photo_id: int,
                        db: Session = Depends(get_db),
                        current_user: User = Depends(repository_users.get_current_user)):
-    deleted_photo = await repository_photos.delete_photo_by_id(
+
+    current_user_role = await repository_users.get_user_role(user_id=current_user.id, db=db)
+
+    photo = await repository_photos.get_photo_by_photo_id(
         photo_id=photo_id,
-        current_user=current_user,
-        db=db)
-    if deleted_photo is None:
+        db=db
+        )
+
+    if photo is None:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Photo not found")
+
+    if current_user_role.name == 'admin' or photo.created_by == current_user.id:
+        deleted_photo = await repository_photos.delete_photo_by_id(
+            photo_id=photo_id,
+            db=db)
+    else:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Forbidden, only the owner or admin can delete the photo")
+
     return PhotoResponse(
         id=deleted_photo.id,
         url=deleted_photo.url,
